requirements.txt
annotated-types==0.7.0
argcomplete==3.5.3
asgiref==3.8.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.36.7
botocore==1.36.7
Brlapi==0.8.5
Brotli==1.1.0
cachetools==5.5.1
certifi==2023.5.7
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
Django==5.1.6
dnf==4.22.0
fedora-third-party==0.10
file-magic==0.4.0
fonttools==4.56.0
fros==1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
humanize==4.9.0
idna==3.7
iso639==0.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
langtable==0.0.68
libcomps==0.1.21
libdnf==0.73.4
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
Markdown==3.7
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.0
mutagen==1.47.0
nftables==0.1
nltk==3.9.1
numpy==2.2.3
olefile==0.47
packaging==24.2
pandas==2.2.3
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
ply==3.11
productmd==1.43
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.21.0
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==24.2.1
pyparsing==3.2.1
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.20.0
rsa==4.9
s3transfer==0.11.2
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.8.2
soupsieve==2.6
sqlparse==0.5.3
systemd-python==235
Tempita==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
websockets==12.0
xkbregistry==0.3
yt-dlp==2025.1.26
